Move PyramidalService and PyramidalPreprocessor to pyramid.fiji.plugins
Groups the @Plugin-annotated classes in the fiji layer under their own
subpackage, mirroring the plugins package already used on the fiji-ui side.
